% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_character.R
\name{dis_character}
\alias{dis_character}
\title{Validate Character Arguments}
\usage{
dis_character(x, valid = NULL, null_valid = TRUE, empty_valid = FALSE,
    scalar = TRUE, param = NULL, call = NULL, fact_check = "global")
}
\arguments{
\item{x}{Required object; a parameter argument to test.}

\item{valid}{Optional character scalar or vector; valid values for the parameter.}

\item{null_valid}{Required logical scalar; whether the parameter can be \code{NULL}.
If \code{FALSE}, the function will throw an error if \code{x} is \code{NULL}.
Default is \code{TRUE}.}

\item{empty_valid}{Required logical scalar; whether the parameter can be an empty
string. If \code{FALSE} (default), the function will throw an error if \code{x}
is an empty string. If \code{TRUE}, the function will allow an empty string.}

\item{scalar}{Required logical scalar; whether the parameter must be a scalar.
If \code{TRUE} (default), the function will throw an error if \code{x} is
not a scalar.}

\item{param}{Optional character scalar; the parameter name. If \code{NULL} (default),
the function will attempt to determine the parameter name from the calling
environment. If nesting functions, it is recommended to provide the parameter
name to ensure the correct parameter is referenced using \code{rlang::caller_arg()}.}

\item{call}{Optional environment; the environment in which the function was called.
If \code{NULL} (default), the function will attempt to determine the calling
environment. If nesting functions, it is recommended to provide the calling
environment to ensure the correct environment is referenced using
\code{rlang::caller_env()}.}

\item{fact_check}{Required character scalar; whether to override fact checking
environment setting. If \code{"global"} (default), \code{dis_character} will
follow the global setting. If \code{"always"}, \code{dis_character} will
ignore any global setting and will always check \code{x}. This argument is
primarily intended for Shiny developers who wish to use \code{disputeR} in
modules. See the vignette on \code{vignette("developing", package = "disputeR")}
for details on how to use this function.}
}
\value{
This function will return either \code{TRUE} (if \code{x} passes
    all validation checks) or \code{FALSE} (if the validation checks are
    skipped). If \code{x} fails validation checks, an error message will
    be returned. Note that, if the input is \code{NULL} and \code{null_valid}
    is set to \code{TRUE}, the detailed unit tests are skipped and the
    function will return \code{TRUE}.
}
\description{
This function runs standard unit tests on character parameters for
    functions.
}
\details{
See the vignette on \code{vignette("developing", package = "disputeR")}
    for details about internal validation of arguments for this function.
}
\examples{
# create example function that uses dis_character()
example <- function(x){

  ## check inputs with disputeR
  dis_not_missing(.f = rlang::is_missing(x))
  dis_character(x, null_valid = FALSE)

  ## modify string
  out <- paste0(x, " rules!")

  ## return output
  return(out)

}

# test example function
example(x = "R")

}
