% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_fact_check.R
\name{dis_checker}
\alias{dis_checker}
\title{Check and Set fact_check Within a Function}
\usage{
dis_checker(fact_check, call = rlang::caller_env())
}
\arguments{
\item{fact_check}{Required character scalar; one of either \code{"global"} or
\code{"always"}, passed from the parent function's \code{fact_check}
parameter.}

\item{call}{Required environment; the environment in which the function was called.
If \code{NULL} (default), the function will attempt to determine the calling
environment. If nesting functions, it is recommended to provide the calling
environment to ensure the correct environment is referenced using
\code{rlang::caller_env()}.}
}
\value{
Either \code{TRUE} if fact checking should be completed or \code{FALSE}
    if it should be skipped. If the value of \code{fact_check} is not either
    \code{"global"} or \code{"always"}, an error will be returned.
}
\description{
This is designed to be used within functions to check the
    status of the \code{DISPUTER_FACT_CHECK} environment variable as well
    as the \code{fact_check} argument. Both are evaluated and the output
    can be used to determine the execution path for the remainder of the
    function.
}
\details{
Since this function is meant to be used prior to most other checks,
    internal validation of its arguments cannot be skipped. See the vignette on
    \code{vignette("developing", package = "disputeR")} for details.
}
\examples{
dis_checker(fact_check = "global")

}
