% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_fact_check.R
\name{dis_dev_check}
\alias{dis_dev_check}
\title{Set the Developer Check Environment Variable}
\usage{
dis_dev_check(dev_check = FALSE)
}
\arguments{
\item{dev_check}{Required logical scalar; should parameter checks on
\code{disputeR} functions be conducted? If \code{FALSE} (default), all
checks for \code{disputeR} function parameters will be skipped. This is
the ideal situation for end users who are using a function with properly
configured \code{disputeR} functions within it. For developers who are
checking the configuration of \code{disputeR} functions, setting
\code{dev_check} to \code{TRUE} can aid with troubleshooting. See the
vignette on \code{vignette("developing", package = "disputeR")}
for details on how to use this function.}
}
\value{
This function writes to your session's environment variables, but
    does not explicitly return any value.
}
\description{
The \code{DISPUTER_DEV_CHECK} environment variable is used
    determine whether parameter checks on \code{disputeR} functions are run
    during a session.
}
\details{
Since this function is meant to be used interactively, internal validation
    of its arguments cannot be skipped. See the vignette on
    \code{vignette("developing", package = "disputeR")} for details.
}
