% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_dir.R
\name{dis_dir}
\alias{dis_dir}
\title{Validate Directory Arguments}
\usage{
dis_dir(x, null_valid = TRUE, error = TRUE, param = NULL,
    call = NULL, fact_check = "global")
}
\arguments{
\item{x}{Required object; a parameter argument to test.}

\item{null_valid}{Required logical scalar; whether the parameter can be \code{NULL}.
If \code{FALSE}, the function will throw an error if \code{x} is \code{NULL}.
Default is \code{TRUE}.}

\item{error}{Required logical scalar; whether the function should error if the
path is not found. If \code{TRUE} (default), the function will throw an
error if \code{x} is not found. If \code{FALSE}, the function will create
a directory at the given path instead of erroring.}

\item{param}{Optional character scalar; the parameter name. If \code{NULL} (default),
the function will attempt to determine the parameter name from the calling
environment. If nesting functions, it is recommended to provide the parameter
name to ensure the correct parameter is referenced using \code{rlang::caller_arg()}.}

\item{call}{Optional environment; the environment in which the function was called.
If \code{NULL} (default), the function will attempt to determine the calling
environment. If nesting functions, it is recommended to provide the calling
environment to ensure the correct environment is referenced using
\code{rlang::caller_env()}.}

\item{fact_check}{Required character scalar; whether to override fact checking
environment setting. If \code{"global"} (default), \code{dis_character} will
follow the global setting. If \code{"always"}, \code{dis_character} will
ignore any global setting and will always check \code{x}. This argument is
primarily intended for Shiny developers who wish to use \code{disputeR} in
modules. See the vignette on \code{vignette("developing", package = "disputeR")}
for details on how to use this function.}
}
\value{
This function will return either \code{TRUE} (if the input passes
    all validation checks). If it does not pass all validation tests, it will
    return an error message if \code{error = TRUE} or a message about the
    creation of a directory if \code{error = FALSE}. Note that, if the input
    is \code{NULL} and \code{null_valid} is set to \code{TRUE}, the detailed
    unit tests are skipped and the function will return \code{TRUE}.
}
\description{
This function runs standard unit tests on a character argument
    that should resolve to a specific directory.
}
\details{
See the vignette on \code{vignette("developing", package = "disputeR")}
    for details about internal validation of arguments for this function.
}
