% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_fact_check.R
\name{dis_fact_check}
\alias{dis_fact_check}
\title{Set the Fact Check Environment Variable}
\usage{
dis_fact_check(fact_check = TRUE)
}
\arguments{
\item{fact_check}{Required logical scalar; should \code{disputeR} fact checking
be conducted? If \code{TRUE} (default), all parameter checks within a project or
Shiny app will be run. When set to \code{FALSE}, all checks will be skipped
unless they are specifically overridden with a function's \code{fact_check}
argument. See the vignette on \code{vignette("developing", package = "disputeR")}
for details on how to use this function.}
}
\value{
This function writes to your session's environment variables, but
    does not explicitly return any value.
}
\description{
The \code{DISPUTER_FACT_CHECK} environment variable is used
    determine whether \code{disputeR} checks are run globally within a package
    or Shiny app.
}
\details{
Since this function is meant to be used interactively, internal validation
    of its arguments cannot be skipped. See the vignette on
    \code{vignette("developing", package = "disputeR")} for details.
}
\examples{
dis_fact_check()

}
