% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_indefinite.R
\name{dis_indefinite}
\alias{dis_indefinite}
\title{Determine Whether an Indefinite Article is Needed for a Particular Class}
\usage{
dis_indefinite(x, indefinite_class = NULL)
}
\arguments{
\item{x}{Required object; a parameter argument to test.}

\item{indefinite_class}{Optional character vector; if the class of an object
is listed here, and \code{class(x)} matches this class, the function
will return \code{"an"}.}
}
\value{
If the input is an integer or the class of the input is included in
    \code{indefinite_class}, the function will return \code{"an"}. Otherwise,
    the function will return \code{"a"}.
}
\description{
This function returns the correct indefinite article based on the
   class of the input.
}
\details{
See the vignette on \code{vignette("developing", package = "disputeR")}
    for details about internal validation of arguments for this function. Unlike
    the core functions, \code{dis_param} does not have a \code{fact_check}
    argument. The \code{vignette("developing", package = "disputeR")} vignette
    includes details on how to implement that functionality around \code{dis_param}.
}
\examples{
dis_indefinite(x = "test")
dis_indefinite(x = 2)

}
