% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_msg.R
\name{dis_msg_miss}
\alias{dis_msg_miss}
\title{Default Message for Missing Parameters}
\usage{
dis_msg_miss()
}
\value{
A character string with the default message.
}
\description{
Create the default message used when a parameter is found to
    be missing.
}
\examples{
dis_msg_miss()

}
