% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_msg.R
\name{dis_msg_valid}
\alias{dis_msg_valid}
\title{Default Message for Invalid Parameters}
\usage{
dis_msg_valid()
}
\value{
A character string with the default message.
}
\description{
Create the default message used when a parameter is found to
    be have an invalid value.
}
\examples{
dis_msg_valid()

}
