% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_numeric.R
\name{dis_numeric}
\alias{dis_numeric}
\title{Validate Numeric Arguments}
\usage{
dis_numeric(x, valid = NULL, null_valid = TRUE, na_valid = FALSE,
    nan_valid = FALSE, infinite_valid = FALSE, integer = FALSE, scalar = TRUE,
    param = NULL, call = NULL, fact_check = "global")
}
\arguments{
\item{x}{Required object; a parameter argument to test.}

\item{valid}{Optional numeric scalar or vector; valid values for the parameter.}

\item{null_valid}{Required logical scalar; whether the parameter can be \code{NULL}.
If \code{FALSE}, the function will throw an error if \code{x} is \code{NULL}.
Default is \code{TRUE}.}

\item{na_valid}{Required logical scalar; whether the parameter can be \code{NA}.
If \code{FALSE}, the function will throw an error if \code{x} is \code{NA}.
Default is \code{FALSE}.}

\item{nan_valid}{Required logical scalar; whether the parameter can be \code{NaN}.
If \code{FALSE}, the function will throw an error if \code{x} is \code{NaN}.
Default is \code{FALSE}.}

\item{infinite_valid}{Required logical scalar; whether the parameter can be \code{Inf}
or \code{-Inf}. If \code{FALSE}, the function will throw an error if \code{x}
is infinite. Default is \code{FALSE}.}

\item{integer}{Required logical scalar; whether the parameter must be an integer.
If \code{TRUE}, the function will throw an error if \code{x} is not an integer.
Default is \code{FALSE}.}

\item{scalar}{Required logical scalar; whether the parameter must be a scalar.
If \code{TRUE} (default), the function will throw an error if \code{x} is
not a scalar.}

\item{param}{Optional character scalar; the parameter name. If \code{NULL} (default),
the function will attempt to determine the parameter name from the calling
environment. If nesting functions, it is recommended to provide the parameter
name to ensure the correct parameter is referenced using \code{rlang::caller_arg()}.}

\item{call}{Optional environment; the environment in which the function was called.
If \code{NULL} (default), the function will attempt to determine the calling
environment. If nesting functions, it is recommended to provide the calling
environment to ensure the correct environment is referenced using
\code{rlang::caller_env()}.}

\item{fact_check}{Required character scalar; whether to override fact checking
environment setting. If \code{"global"} (default), \code{dis_character} will
follow the global setting. If \code{"always"}, \code{dis_character} will
ignore any global setting and will always check \code{x}. This argument is
primarily intended for Shiny developers who wish to use \code{disputeR} in
modules. See the vignette on \code{vignette("developing", package = "disputeR")}
for details on how to use this function.}
}
\value{
This function will return either \code{TRUE} (if the input passes
    all validation checks) or an error message. Note that, if the input is \code{NULL},
    \code{NA}, \code{NaN}, or infinite and the appropriate arguments are set
    to \code{TRUE}, the detailed unit tests are skipped and the function will
    return \code{TRUE}.
}
\description{
This function runs standard unit tests on numeric and integer
    parameters for functions.
}
\details{
See the vignette on \code{vignette("developing", package = "disputeR")}
    for details about internal validation of arguments for this function.
}
\examples{
# create example function that uses dis_numeric()
example <- function(x){

  ## check inputs with disputeR
  dis_not_missing(.f = rlang::is_missing(x))
  dis_numeric(x, null_valid = FALSE)

  ## square
  out <- x^2

  ## return output
  return(out)

}

# test example function
example(x = 2)

}
