% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dis_not_missing.R
\name{dis_not_missing}
\alias{dis_not_missing}
\title{Validate Whether Parameter is Missing}
\usage{
dis_not_missing(.f, param = "x", call = rlang::caller_env())
}
\arguments{
\item{.f}{Required logical scalar; the output of \code{missing(x))} or
\code{rlang::is_missing(x)}, substituting the parameter name passed to
either function as needed.}

\item{param}{Required character scalar; the parameter name. The default uses
the variable name \code{x}, but this value can be set to a different
static character scalar. Alternatively, you can use \code{rlang::caller_arg()}
to pull the parameter name passed by the function \code{dis_not_missing()}
is nested within. Regardless of the approach taken here, the resulting
character scalar should match the object name passed to \code{.f}.}

\item{call}{Required environment; the environment in which the function was called.
If nesting \code{dis_not_missing} within another function, it is recommended to
provide the calling environment to ensure the correct environment is referenced
using \code{rlang::caller_env()} (default). If there are multiple levels of
nesting, the \code{call} argument can be used to pass an environment created
in the outermost function.}
}
\value{
This function will return either \code{TRUE} (if the input passes
    all validation checks) or an error message.
}
\description{
This function returns a standard error message for missing parameters.
}
\details{
See the vignette on \code{vignette("performance", package = "disputeR")}
    for details about how to skip internal validation of arguments for this
    function.
}
\examples{
# create example function that uses dis_param()
example <- function(x){
  dis_not_missing(.f = rlang::is_missing(x))
}

# example function
example(x = "test")

# create example function that uses dis_param() with different variable name
example <- function(var){
  dis_not_missing(.f = rlang::is_missing(var), param = "var")
}

# example function
example(var = "test")

}
